feat: Implement authentication and user management
- Added Spring Security dependency for password encryption.
- Created AuthController for user login handling.
- Introduced LoginRequest DTO for login data transfer.
- Refactored Cliente, Funcionario, and Proprietario models to extend a new Usuario base class.
- Implemented password hashing and verification in PasswordUtil.
- Updated ClienteService, FuncionarioService, and ProprietarioService to handle password encryption on save and update.
- Added UsuarioController for user retrieval and management.
- Enhanced repository interfaces to include email and CPF checks.
- Removed unused JavaScript code from frontend.

+++ backend/src/main/java/com/github/peu06/agendamento_api/service/AgendamentoService.java
@@ -41,7 +41,6 @@ public class AgendamentoService {
 
         if (agenda.getInicio().isBefore(LocalDateTime.now())){
             agenda.setStatus(Agenda.Status.HORARIO_PASSADO);
-            agendaRepository.save(agenda);
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Não é possível agendar horário passado");
         }
 
@@ -94,14 +93,18 @@ public class AgendamentoService {
         return agendamentoRepository.save(agendamentoExistente);
     }
 
+    @Transactional
     public void delete(Long id) {
 
         Agendamento agendamento = getById(id);
 
         Agenda agenda = agendamento.getAgenda();
-        agenda.setStatus(Agenda.Status.DISPONIVEL);
 
-        agendaRepository.save(agenda);
+        if (agenda.getInicio().isBefore(LocalDateTime.now().plusHours(1))){
+            throw new RuntimeException("Cancelamento deve ser feito com 1 hora de antecedência");
+        }
+
+        agenda.setStatus(Agenda.Status.DISPONIVEL);
 
         agendamentoRepository.delete(agendamento);
     }
